requirements.txt
flask
flask-cors
opencv-python
mediapipe
numpy
groq
pymongo
python-dotenv